app/lib/multimedia/handlers/pdf-handler.ts
/**
 * PDF Handler
 * 
 * Handles PDF processing and image extraction:
 * 1. PDF parsing
 * 2. Image extraction
 * 3. Metadata extraction
 * 4. Page processing
 * 5. Image optimization
 */

import { PDFDocument, PDFPage } from 'pdf-lib';
import { createImageHandler, type ProcessedImage } from './image-handler';
import { metricsCollector, MetricType, MetricCategory } from '../../../lib/ml/code-understanding/telemetry/metrics-collector';
import { performanceTracker, OperationType } from '../../../lib/ml/code-understanding/telemetry/performance-tracker';

export interface PDFMetadata {
  title?: string;
  author?: string;
  subject?: string;
  keywords?: string[];
  creator?: string;
  producer?: string;
  creationDate?: Date;
  modificationDate?: Date;
  pageCount: number;
}

export interface PDFImageInfo {
  pageNumber: number;
  width: number;
  height: number;
  bitsPerComponent: number;
  colorSpace: string;
}

export interface ExtractedImage {
  imageInfo: PDFImageInfo;
  buffer: Buffer;
}

export interface PDFProcessingOptions {
  pageRange?: {
    start?: number;
    end?: number;
  };
  imageQuality?: number;
  minImageSize?: number;
  maxImageSize?: number;
  extractMetadata?: boolean;
}

interface PDFImageObject {
  width: number;
  height: number;
  bitsPerComponent: number;
  colorSpace: string;
  buffer: Buffer;
}

export class PDFHandler {
  private static instance: PDFHandler;
  private imageHandler = createImageHandler();

  private constructor() {
    this.initializeMetrics();
  }

  static getInstance(): PDFHandler {
    if (!PDFHandler.instance) {
      PDFHandler.instance = new PDFHandler();
    }
    return PDFHandler.instance;
  }

  /**
   * Extract images from PDF
   */
  async extractImages(
    pdfBuffer: Buffer | ArrayBuffer,
    options: PDFProcessingOptions = {}
  ): Promise<ProcessedImage[]> {
    const operationId = performanceTracker.startOperation({
      component: 'PDFHandler',
      operation: OperationType.RESOURCE_CLEANUP
    });

    try {
      const pdfDoc = await PDFDocument.load(pdfBuffer);
      const metadata = await this.extractMetadata(pdfDoc);
      const extractedImages: ExtractedImage[] = [];

      const startPage = options.pageRange?.start || 1;
      const endPage = options.pageRange?.end || metadata.pageCount;

      // Process each page in the specified range
      for (let pageNum = startPage; pageNum <= endPage; pageNum++) {
        const page = pdfDoc.getPage(pageNum - 1);
        const pageImages = await this.extractImagesFromPage(page, pageNum);
        extractedImages.push(...pageImages);
      }

      // Filter and process images
      const processedImages = await this.processExtractedImages(
        extractedImages,
        options
      );

      // Record metrics
      this.recordMetrics(metadata, extractedImages, processedImages);

      return processedImages;
    } finally {
      performanceTracker.endOperation(operationId, {
        component: 'PDFHandler',
        operation: OperationType.RESOURCE_CLEANUP
      });
    }
  }

  /**
   * Extract metadata from PDF
   */
  private async extractMetadata(pdfDoc: PDFDocument): Promise<PDFMetadata> {
    const info = pdfDoc.getDocumentInfo();
    
    return {
      title: info.title || undefined,
      author: info.author || undefined,
      subject: info.subject || undefined,
      keywords: info.keywords?.split(',').map((k: string) => k.trim()),
      creator: info.creator || undefined,
      producer: info.producer || undefined,
      creationDate: info.creationDate,
      modificationDate: info.modificationDate,
      pageCount: pdfDoc.getPageCount()
    };
  }

  /**
   * Extract images from a single page
   */
  private async extractImagesFromPage(
    page: PDFPage,
    pageNumber: number
  ): Promise<ExtractedImage[]> {
    const images: ExtractedImage[] = [];
    
    // Implementation note: The actual image extraction would be more complex
    // and would involve parsing the page's content stream to find and extract
    // image objects. This is a simplified version.
    
    // For each image object in the page...
    const imageObjects: PDFImageObject[] = []; // Would be populated by parsing page content
    for (const imgObj of imageObjects) {
      const imageInfo: PDFImageInfo = {
        pageNumber,
        width: imgObj.width,
        height: imgObj.height,
        bitsPerComponent: imgObj.bitsPerComponent,
        colorSpace: imgObj.colorSpace
      };

      images.push({
        imageInfo,
        buffer: imgObj.buffer
      });
    }

    return images;
  }

  /**
   * Process extracted images
   */
  private async processExtractedImages(
    extractedImages: ExtractedImage[],
    options: PDFProcessingOptions
  ): Promise<ProcessedImage[]> {
    const processedImages: ProcessedImage[] = [];

    for (const { buffer, imageInfo } of extractedImages) {
      // Filter by size if specified
      if (options.minImageSize && buffer.length < options.minImageSize) {
        continue;
      }
      if (options.maxImageSize && buffer.length > options.maxImageSize) {
        continue;
      }

      // Process image using ImageHandler
      const processed = await this.imageHandler.processImage(buffer, `page_${imageInfo.pageNumber}_image.png`, {
        quality: options.imageQuality
      });

      processedImages.push(processed);
    }

    return processedImages;
  }

  private recordMetrics(
    metadata: PDFMetadata,
    extractedImages: ExtractedImage[],
    processedImages: ProcessedImage[]
  ): void {
    metricsCollector.record('pdf_pages_processed', metadata.pageCount);
    metricsCollector.record('pdf_images_extracted', extractedImages.length);
    metricsCollector.record('pdf_images_processed', processedImages.length);

    const totalOriginalSize = extractedImages.reduce(
      (sum, img) => sum + img.buffer.length,
      0
    );
    const totalProcessedSize = processedImages.reduce(
      (sum, img) => sum + img.metadata.size,
      0
    );

    metricsCollector.record('pdf_images_original_size', totalOriginalSize);
    metricsCollector.record('pdf_images_processed_size', totalProcessedSize);
  }

  private initializeMetrics(): void {
    metricsCollector.registerMetric({
      name: 'pdf_pages_processed',
      type: MetricType.COUNTER,
      category: MetricCategory.SYSTEM,
      description: 'Number of PDF pages processed'
    });

    metricsCollector.registerMetric({
      name: 'pdf_images_extracted',
      type: MetricType.COUNTER,
      category: MetricCategory.SYSTEM,
      description: 'Number of images extracted from PDFs'
    });

    metricsCollector.registerMetric({
      name: 'pdf_images_processed',
      type: MetricType.COUNTER,
      category: MetricCategory.SYSTEM,
      description: 'Number of PDF images successfully processed'
    });

    metricsCollector.registerMetric({
      name: 'pdf_images_original_size',
      type: MetricType.GAUGE,
      category: MetricCategory.SYSTEM,
      description: 'Total size of original images in bytes'
    });

    metricsCollector.registerMetric({
      name: 'pdf_images_processed_size',
      type: MetricType.GAUGE,
      category: MetricCategory.SYSTEM,
      description: 'Total size of processed images in bytes'
    });
  }
}

// Export factory function
export function createPDFHandler(): PDFHandler {
  return PDFHandler.getInstance();
}
